truthguard_dev/requirements.txt
Flask>=2.3.2
Flask-Cors>=4.0.0
firebase-admin>=6.2.0
scikit-learn>=1.3.0
pandas>=2.0.3
numpy>=1.24.4
nltk>=3.8.1
python-dateutil>=2.8.2
requests>=2.31.0
beautifulsoup4>=4.12.2
bs4>=0.0.1
langdetect>=1.0.9
pyspellchecker>=0.6.2